Run the runtime at native bf16 and delete the fp32 upcast path
权重原生 bf16,vllm-omni 参考实现同为 bf16 口径,而实测官方 Qwen3 自身
fp32/bf16 贪心序列在 0–8 token 内即分歧——bf16 漂移是 dtype 固有,任何
bf16 runtime 都无法"对齐 fp32",该门禁从前提上就不成立。此前把 bf16 权重
零填充到 fp32 运行不恢复任何信息,只让权重流量与 KV 显存翻倍。据此把
runtime 收敛为 bf16 原生单一精度:删除 static_graph_decode.py(fp32 专用
分桶图捕获)、--thinker-dtype/--no-graph-decode/--fp32 开关与 fp32 默认值;
Thinker decode 统一走 sparkinfer paged CUDA graph(对 bf16 eager 61/61
token 逐位一致,20.7 vs 175.9ms/token,8.5x);门禁收敛为同精度
graph/eager parity + 端到端信号门禁。端到端 realtime x3:回复与信号门禁
全过,峰值显存 50.5→28.3GiB,chunk 决策 19–175ms。

Constraint: 权重原生 bf16,运行时不做反量化、不提高运行时精度(用户指令)
Constraint: 质量门禁必须可执行:同内核族 graph/eager 贪心逐 token 一致 + fixture 信号门禁
Rejected: 继续以"对齐 fp32"为 bf16 上线门槛 | 官方自身 fp32/bf16 亦分歧,门禁不可达且与 vllm-omni 口径相悖
Rejected: 保留 fp32 双轨 | 双轨使 runtime 复杂化,且 fp32 无独立质量证据
Confidence: high
Scope-risk: broad
Reversibility: messy(删除面较大,但 git 历史完整保留 fp32 路径与 48/48、121/121 parity 存档)
Directive: 新增模型路径一律 bf16 原生;若未来需要结构级验证,从 git 历史恢复 fp32 parity 工具,不要在 runtime 内重建
Tested: p1_graph_decode_check 61/61 PASS 与分离计时; pytest -q 166 passed; ruff; p1_duplex_loop realtime x3(回复、信号门禁、显存)
Not-tested: 长会话 bf16 漂移累积的听感评估; 真实 vLLM-omni 跨 runtime 对照

+++ channellm/engine/blocks.py
@@ -129,7 +129,7 @@ class TorchStaticKV:
         head_dim: int,
         *,
         device: torch.device | str = "cpu",
-        dtype: torch.dtype = torch.float32,
+        dtype: torch.dtype = torch.bfloat16,
     ) -> None:
         if min(num_layers, max_seq_len, num_kv_heads, head_dim) <= 0:
             raise ValueError("static KV dimensions must be positive")
