+++ channellm/engine/thinker.py
@@ -143,7 +143,7 @@ class Thinker(nn.Module):
         self,
         config: ThinkerConfig,
         device: str | torch.device = "cpu",
-        dtype: torch.dtype = torch.float32,
+        dtype: torch.dtype = torch.bfloat16,
     ) -> None:
         super().__init__()
         self.config = config
